2.12.0

+++ bp-better-messages/inc/functions.php
@@ -3308,13 +3308,13 @@ if ( !class_exists( 'Better_Messages_Functions' ) ):
             $forced_user_id = apply_filters( 'better_messages_forced_current_user_id', null );
 
             if( $forced_user_id !== null ){
-                return $forced_user_id;
+                return (int) $forced_user_id;
             }
 
             if( is_user_logged_in() ){
-                return apply_filters('better_messages_logged_in_user_id', get_current_user_id() );
+                return (int) apply_filters('better_messages_logged_in_user_id', get_current_user_id() );
             } else {
-                return apply_filters('better_messages_guest_user_id', 0);
+                return (int) apply_filters('better_messages_guest_user_id', 0);
             }
         }
